spectrogramxApp/requirements.txt
pysctp
matplotlib
numpy
Pillow
ricsdl
